TropicalGeometry: tropical_variety, moved from traversal.jl to variety.jl + storing data in attributes now
